Scale factors are now applied to streaming data so that they are in m, m/s, m/s^2 and radians.

+++ src/com/jpdelacroix/sphero/SpheroApplication.java
@@ -46,7 +46,7 @@ public class SpheroApplication implements Runnable {
 			if (packet.isAsynchronous()) {
 				System.out.println("");
 				System.out.println(packet);
-				HashMap<String, Integer> data = ((SpheroAsynchronousPacket) packet).parseDataWithOptions(options);
+				HashMap<String, Double> data = ((SpheroAsynchronousPacket) packet).parseDataWithOptions(options);
 				for (String name : data.keySet()) {
 					System.out.println(name + " : " + data.get(name));
 				}
